+++ src/ScottPlot/Markers.cs
@@ -1,10 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Drawing;
-using System.Drawing.Drawing2D;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Runtime.InteropServices;
 
 namespace ScottPlot
@@ -40,6 +36,14 @@ namespace ScottPlot
             Pen pen = new Pen(color);
 
             Brush brush = new SolidBrush(color);
+
+            // Improve marker vs. line alignment on Linux and MacOS
+            // https://github.com/ScottPlot/ScottPlot/issues/340
+            // https://github.com/ScottPlot/ScottPlot/pull/1660
+            if (!RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+            {
+                pixelLocation = new PointF(pixelLocation.X + .5f, pixelLocation.Y);
+            }
 
             PointF corner1 = new PointF(pixelLocation.X - size / 2, pixelLocation.Y - size / 2);
             PointF corner2 = new PointF(pixelLocation.X + size / 2, pixelLocation.Y + size / 2);
@@ -186,225 +190,9 @@ namespace ScottPlot
 
         public static void DrawMarkers(Graphics gfx, ICollection<PointF> pixelLocations, MarkerShape shape, float size, Color color)
         {
-            if (size == 0 || shape == MarkerShape.none)
-                return;
-
-            Pen pen = new Pen(color);
-
-            Brush brush = new SolidBrush(color);
-
-            float halfsize = size / 2;
-            float quartersize = size / 4;
-            float halfsqrt3size = (float)0.866 * size;
-            float halfsqrt2size = (float)0.707 * size;
-            float quartersqrt2size = halfsqrt2size / 2;
-            float quartersqrt3size = halfsqrt3size / 2;
-
-            // adjust marker offset to improve rendering on Linux and MacOS
-            float markerOffsetX = (RuntimeInformation.IsOSPlatform(OSPlatform.Windows)) ? 0 : 1;
-            switch (shape)
+            foreach (PointF pt in pixelLocations)
             {
-                case MarkerShape.filledCircle:
-                    foreach (PointF point in pixelLocations)
-                    {
-                        gfx.FillEllipse(brush: brush, x: point.X - halfsize + markerOffsetX, y: point.Y - halfsize, width: size, height: size);
-                    }
-
-                    break;
-                case MarkerShape.openCircle:
-                    foreach (PointF point in pixelLocations)
-                    {
-                        gfx.DrawEllipse(pen: pen, x: point.X - halfsize + markerOffsetX, y: point.Y - halfsize, width: size, height: size);
-                    }
-
-                    break;
-                case MarkerShape.filledSquare:
-                    foreach (PointF point in pixelLocations)
-                    {
-                        gfx.FillRectangle(brush: brush, x: point.X - halfsize + markerOffsetX, y: point.Y - halfsize, width: size, height: size);
-                    }
-
-                    break;
-                case MarkerShape.openSquare:
-                    foreach (PointF point in pixelLocations)
-                    {
-                        gfx.DrawRectangle(pen: pen, x: point.X - halfsize + markerOffsetX, y: point.Y - halfsize, width: size, height: size);
-                    }
-
-                    break;
-                case MarkerShape.filledDiamond:
-                    foreach (PointF point in pixelLocations)
-                    {
-                        // Create points that define polygon.
-                        PointF point1 = new PointF(point.X + markerOffsetX, point.Y + halfsize);
-                        PointF point2 = new PointF(point.X + markerOffsetX - halfsize, point.Y);
-                        PointF point3 = new PointF(point.X + markerOffsetX, point.Y - halfsize);
-                        PointF point4 = new PointF(point.X + markerOffsetX + halfsize, point.Y);
-
-                        PointF[] curvePoints = { point1, point2, point3, point4 };
-
-                        //Fill polygon to screen
-                        gfx.FillPolygon(brush, curvePoints);
-                    }
-
-                    break;
-                case MarkerShape.openDiamond:
-                    foreach (PointF point in pixelLocations)
-                    {
-                        // Create points that define polygon.
-                        PointF point5 = new PointF(point.X + markerOffsetX, point.Y + halfsize);
-                        PointF point6 = new PointF(point.X + markerOffsetX - halfsize, point.Y);
-                        PointF point7 = new PointF(point.X + markerOffsetX, point.Y - halfsize);
-                        PointF point8 = new PointF(point.X + markerOffsetX + halfsize, point.Y);
-
-                        PointF[] curvePoints2 = { point5, point6, point7, point8 };
-
-                        //Draw polygon to screen
-                        gfx.DrawPolygon(pen, curvePoints2);
-                    }
-
-                    break;
-                case MarkerShape.asterisk:
-                    //Font drawFont = new Font("CourierNew", size * 3);
-                    //SizeF textSize = Drawing.GDI.MeasureString(gfx, "*", drawFont);
-                    //float halfWidth = textSize.Width / 2;
-                    //float quarterHeight = textSize.Height / 4;
-                    foreach (PointF point in pixelLocations)
-                    {
-                        //gfx.DrawString("*", drawFont, brush, point.X - halfWidth, point.Y - quarterHeight);
-                        // horizontal line of the *
-                        gfx.DrawLine(pen, point.X + markerOffsetX - halfsize, point.Y, point.X + markerOffsetX + halfsize, point.Y);
-                        // vertical line of the *
-                        gfx.DrawLine(pen, point.X + markerOffsetX, point.Y - halfsize, point.X + markerOffsetX, point.Y + halfsize);
-                        // bottom-left / top-right line of the *
-                        gfx.DrawLine(pen, point.X + markerOffsetX - quartersqrt2size, point.Y - quartersqrt2size, point.X + markerOffsetX + quartersqrt2size, point.Y + quartersqrt2size);
-                        // top-left / bottom-right line of the *
-                        gfx.DrawLine(pen, point.X + markerOffsetX - quartersqrt2size, point.Y + quartersqrt2size, point.X + markerOffsetX + quartersqrt2size, point.Y - quartersqrt2size);
-                    }
-
-                    break;
-                case MarkerShape.hashTag:
-                    Font drawFont2 = new Font("CourierNew", size * 2);
-                    SizeF textSize2 = Drawing.GDI.MeasureString(gfx, "#", drawFont2);
-                    float halfWidth2 = textSize2.Width / 2;
-                    float quarterHeight2 = textSize2.Height / 4;
-                    foreach (PointF point in pixelLocations)
-                    {
-                        gfx.DrawString("#", drawFont2, brush, point.X - halfWidth2, point.Y - quarterHeight2);
-                    }
-
-                    break;
-                case MarkerShape.cross:
-                    //Font drawFont3 = new Font("CourierNew", size * 2);
-                    //SizeF textSize3 = Drawing.GDI.MeasureString(gfx, "+", drawFont3);
-                    //float halfWidth3 = textSize3.Width / (5 / 2);
-                    //float quarterHeight3 = textSize3.Height / 2;
-                    foreach (PointF point in pixelLocations)
-                    {
-                        //gfx.DrawString("+", drawFont3, brush, point.X - halfWidth3, point.Y - quarterHeight3);
-                        // horizontal line of the +
-                        gfx.DrawLine(pen, point.X + markerOffsetX - halfsize, point.Y, point.X + markerOffsetX + halfsize, point.Y);
-                        // vertical line of the +
-                        gfx.DrawLine(pen, point.X + markerOffsetX, point.Y - halfsize, point.X + markerOffsetX, point.Y + halfsize);
-                    }
-
-                    break;
-                case MarkerShape.eks:
-                    //Font drawFont4 = new Font("CourierNew", size * 2);
-                    //SizeF textSize4 = Drawing.GDI.MeasureString(gfx, "x", drawFont4);
-                    //float halfWidth4 = textSize4.Width / (5 / 2);
-                    //float quarterHeight4 = textSize4.Height / 2;
-                    foreach (PointF point in pixelLocations)
-                    {
-                        //gfx.DrawString("x", drawFont4, brush, point.X - halfWidth4, point.Y - quarterHeight4);
-                        // bottom-left / top-right line of the x
-                        gfx.DrawLine(pen, point.X + markerOffsetX - quartersqrt2size, point.Y - quartersqrt2size, point.X + markerOffsetX + quartersqrt2size, point.Y + quartersqrt2size);
-                        // top-left / bottom-right line of the x
-                        gfx.DrawLine(pen, point.X + markerOffsetX - quartersqrt2size, point.Y + quartersqrt2size, point.X + markerOffsetX + quartersqrt2size, point.Y - quartersqrt2size);
-                    }
-
-                    break;
-                case MarkerShape.verticalBar:
-                    //Font drawFont5 = new Font("CourierNew", size * 2);
-                    //SizeF textSize5 = Drawing.GDI.MeasureString(gfx, "|", drawFont5);
-                    //float halfWidth5 = textSize5.Width / (5 / 2);
-                    //float quarterHeight5 = textSize5.Height / 2;
-                    foreach (PointF point in pixelLocations)
-                    {
-                        //gfx.DrawString("|", drawFont5, brush, point.X - halfWidth5, point.Y - quarterHeight5);
-                        // vertical line of the *
-                        gfx.DrawLine(pen, point.X + markerOffsetX, point.Y - halfsize, point.X + markerOffsetX, point.Y + halfsize);
-                    }
-
-                    break;
-                case MarkerShape.triUp:
-                    foreach (PointF point in pixelLocations)
-                    {
-                        // Create points that define polygon.
-                        PointF point9 = new PointF(point.X + markerOffsetX, point.Y - halfsize);
-                        PointF point10 = new PointF(point.X + markerOffsetX, point.Y);
-                        PointF point11 = new PointF(point.X + markerOffsetX - quartersqrt3size, point.Y + quartersize);
-                        PointF point12 = new PointF(point.X + markerOffsetX, point.Y);
-                        PointF point13 = new PointF(point.X + markerOffsetX + quartersqrt3size, point.Y + quartersize);
-
-                        PointF[] curvePoints3 = { point12, point9, point10, point11, point12, point13 };
-
-                        //Draw polygon to screen
-                        gfx.DrawPolygon(pen, curvePoints3);
-                    }
-
-                    break;
-                case MarkerShape.triDown:
-                    foreach (PointF point in pixelLocations)
-                    {
-                        // Create points that define polygon.
-                        PointF point14 = new PointF(point.X + markerOffsetX, point.Y + halfsize);
-                        PointF point15 = new PointF(point.X + markerOffsetX, point.Y);
-                        PointF point16 = new PointF(point.X + markerOffsetX - quartersqrt3size, point.Y - quartersize);
-                        PointF point17 = new PointF(point.X + markerOffsetX, point.Y);
-                        PointF point18 = new PointF(point.X + markerOffsetX + quartersqrt3size, point.Y - quartersize);
-
-                        PointF[] curvePoints4 = { point17, point14, point15, point16, point17, point18 };
-
-                        //Draw polygon to screen
-                        gfx.DrawPolygon(pen, curvePoints4);
-                    }
-
-                    break;
-                case MarkerShape.filledTriangle:
-                    foreach (PointF point in pixelLocations)
-                    {
-                        // Create points that define polygon.
-                        PointF point19 = new PointF(point.X + markerOffsetX, point.Y - halfsize);
-                        PointF point20 = new PointF(point.X + markerOffsetX - quartersqrt3size, point.Y + quartersize);
-                        PointF point21 = new PointF(point.X + markerOffsetX + quartersqrt3size, point.Y + quartersize);
-
-                        PointF[] curvePoints5 = { point19, point20, point21 };
-
-                        //Draw polygon to screen
-                        gfx.FillPolygon(brush, curvePoints5);
-                    }
-
-                    break;
-                case MarkerShape.openTriangle:
-                    foreach (PointF point in pixelLocations)
-                    {
-                        // Create points that define polygon.
-                        PointF point22 = new PointF(point.X + markerOffsetX, point.Y - halfsize);
-                        PointF point23 = new PointF(point.X + markerOffsetX - quartersqrt3size, point.Y + quartersize);
-                        PointF point24 = new PointF(point.X + markerOffsetX + quartersqrt3size, point.Y + quartersize);
-
-                        PointF[] curvePoints6 = { point22, point23, point24 };
-
-                        //Draw polygon to screen
-                        gfx.DrawPolygon(pen, curvePoints6);
-                    }
-
-                    break;
-                case MarkerShape.none:
-                    break;
-                default:
-                    throw new NotImplementedException($"unsupported marker type: {shape}");
+                DrawMarker(gfx, pt, shape, size, color);
             }
         }
     }