+++ src/ScottPlot/Plottable/IIsHighlightable.cs
@@ -1,8 +1,0 @@
-﻿namespace ScottPlot.Plottable
-{
-    public interface IIsHighlightable
-    {
-        bool IsHighlighted { get; set; }
-        double HighlightCoefficient { get; set; }
-    }
-}