+++ src/ScottPlot/Plottable/SignalPlotBase.cs
@@ -6,11 +6,10 @@ using System.Drawing;
 using System.Drawing.Drawing2D;
 using System.Linq;
 using System.Linq.Expressions;
-using System.Runtime.InteropServices;
 
 namespace ScottPlot.Plottable
 {
-    public abstract class SignalPlotBase<T> : IPlottable, IHasLine, IHasMarker, IIsHighlightable, IHasPointsGenericX<double, T> where T : struct, IComparable
+    public abstract class SignalPlotBase<T> : IPlottable, IHasLine, IHasMarker, IHighlightable, IHasColor, IHasPointsGenericX<double, T> where T : struct, IComparable
     {
         protected IMinMaxSearchStrategy<T> Strategy = new SegmentedTreeMinMaxSearchStrategy<T>();
         protected bool MaxRenderIndexLowerYSPromise = false;
@@ -21,17 +20,33 @@ namespace ScottPlot.Plottable
         public int YAxisIndex { get; set; } = 0;
         public bool IsVisible { get; set; } = true;
         public bool StepDisplay { get; set; } = false;
-        public float MarkerSize { get; set; } = 5;
+
+        public float _markerSize = 5;
+        public float MarkerSize
+        {
+            get => IsHighlighted ? _markerSize * HighlightCoefficient : _markerSize;
+            set { _markerSize = value; }
+        }
+
         public MarkerShape MarkerShape { get; set; } = MarkerShape.filledCircle;
         public double OffsetX { get; set; } = 0;
         public T OffsetY { get; set; } = default;
-        public double LineWidth { get; set; } = 1;
+
+        private double _lineWidth = 1;
+        public double LineWidth
+        {
+            get => IsHighlighted ? _lineWidth * HighlightCoefficient : _lineWidth;
+            set { _lineWidth = value; }
+        }
+
         public string Label { get; set; } = null;
         public Color Color { get; set; } = Color.Green;
+        public Color LineColor { get => Color; set { Color = value; } }
+        public Color MarkerColor { get => Color; set { Color = value; } }
         public LineStyle LineStyle { get; set; } = LineStyle.Solid;
 
         public bool IsHighlighted { get; set; } = false;
-        public double HighlightCoefficient { get; set; } = 2;
+        public float HighlightCoefficient { get; set; } = 2;
 
         /// <summary>
         /// If enabled, parallel processing will be used to calculate pixel positions for high density datasets.
@@ -359,8 +374,10 @@ namespace ScottPlot.Plottable
                     {
                         ShowMarkersInLegend = true;
 
-                        // draw a marker at each point
-                        MarkerTools.DrawMarkers(gfx, linePoints, MarkerShape, (IsHighlighted ? (float)HighlightCoefficient : 1) * MarkerSize, Color);
+                        foreach (PointF point in linePoints)
+                        {
+                            MarkerTools.DrawMarker(gfx, point, MarkerShape, markerPxDiameter, Color);
+                        }
                     }
                     else
                     {
@@ -728,9 +745,9 @@ namespace ScottPlot.Plottable
                 label = Label,
                 color = Color,
                 lineStyle = LineStyle,
-                lineWidth = (IsHighlighted ? HighlightCoefficient : 1) * LineWidth,
+                lineWidth = LineWidth,
                 markerShape = ShowMarkersInLegend ? MarkerShape.filledCircle : MarkerShape.none,
-                markerSize = ShowMarkersInLegend ? (IsHighlighted ? (float)HighlightCoefficient : 1) * MarkerSize : 0
+                markerSize = ShowMarkersInLegend ? MarkerSize : 0
             };
             return new LegendItem[] { singleLegendItem };
         }
@@ -739,8 +756,8 @@ namespace ScottPlot.Plottable
         {
             using var gfx = GDI.Graphics(bmp, dims, lowQuality);
             using var brush = GDI.Brush(Color);
-            using var penLD = GDI.Pen(Color, (float)((IsHighlighted ? HighlightCoefficient : 1) * LineWidth), LineStyle, true);
-            using var penHD = GDI.Pen(Color, (float)((IsHighlighted ? HighlightCoefficient : 1) * LineWidth), LineStyle.Solid, true);
+            using var penLD = GDI.Pen(Color, (float)LineWidth, LineStyle, true);
+            using var penHD = GDI.Pen(Color, (float)LineWidth, LineStyle.Solid, true);
 
             double dataSpanUnits = _Ys.Length * _SamplePeriod;
             double columnSpanUnits = dims.XSpan / dims.DataWidth;